requirements.txt
python-dotenv==0.9.1
requests==2.28.1
urllib3==1.26.14
